+++ examples/default/variables.tf
@@ -13,3 +13,10 @@ variable "environment" {
   type    = string
   default = null
 }
+
+variable "aws_region" {
+  description = "AWS region."
+
+  type    = string
+  default = "eu-west-1"
+}