+++ lambdas/libs/aws-powertools-util/src/tracer/index.ts
@@ -1,8 +1,6 @@
 import { Tracer, captureLambdaHandler } from '@aws-lambda-powertools/tracer';
 
-const tracer = new Tracer({
-  serviceName: process.env.SERVICE_NAME || 'runners',
-});
+const tracer = new Tracer();
 
 function getTracedAWSV3Client<T>(client: T): T {
   return tracer.captureAWSv3Client(client);