+++ main.tf
@@ -265,6 +265,7 @@ module "runners" {
   enable_userdata                = var.enable_userdata
   enable_user_data_debug_logging = var.enable_user_data_debug_logging_runner
   userdata_template              = var.userdata_template
+  userdata_content               = var.userdata_content
   userdata_pre_install           = var.userdata_pre_install
   userdata_post_install          = var.userdata_post_install
   key_name                       = var.key_name
@@ -288,6 +289,7 @@ module "runners" {
   pool_lambda_reserved_concurrent_executions = var.pool_lambda_reserved_concurrent_executions
 
   ssm_housekeeper = var.runners_ssm_housekeeper
+  ebs_optimized   = var.runners_ebs_optimized
 }
 
 module "runner_binaries" {
@@ -363,4 +365,32 @@ module "ami_housekeeper" {
 
   cleanup_config             = var.ami_housekeeper_cleanup_config
   lambda_schedule_expression = var.ami_housekeeper_lambda_schedule_expression
+}
+
+locals {
+  lambda_instance_termination_watcher = {
+    prefix                    = var.prefix
+    tags                      = local.tags
+    aws_partition             = var.aws_partition
+    architecture              = var.lambda_architecture
+    principals                = var.lambda_principals
+    runtime                   = var.lambda_runtime
+    security_group_ids        = var.lambda_security_group_ids
+    subnet_ids                = var.lambda_subnet_ids
+    log_level                 = var.log_level
+    logging_kms_key_id        = var.logging_kms_key_id
+    logging_retention_in_days = var.logging_retention_in_days
+    role_path                 = var.role_path
+    role_permissions_boundary = var.role_permissions_boundary
+    metrics_namespace         = var.metrics_namespace
+    s3_bucket                 = var.lambda_s3_bucket
+    tracing_config            = var.tracing_config
+  }
+}
+
+module "instance_termination_watcher" {
+  source = "./modules/termination-watcher"
+  count  = var.instance_termination_watcher.enable ? 1 : 0
+
+  config = merge(local.lambda_instance_termination_watcher, var.instance_termination_watcher)
 }