+++ modules/lambda/iam.tf
@@ -1,0 +1,16 @@
+data "aws_iam_policy_document" "lambda_xray" {
+  count = var.lambda.tracing_config.mode != null ? 1 : 0
+  statement {
+    actions = [
+      "xray:BatchGetTraces",
+      "xray:GetTraceSummaries",
+      "xray:PutTelemetryRecords",
+      "xray:PutTraceSegments"
+    ]
+    effect = "Allow"
+    resources = [
+      "*"
+    ]
+    sid = "AllowXRay"
+  }
+}