+++ modules/runners/ssm-housekeeper.tf
@@ -32,7 +32,7 @@ resource "aws_lambda_function" "ssm_housekeeper" {
       ENVIRONMENT                              = var.prefix
       LOG_LEVEL                                = var.log_level
       SSM_CLEANUP_CONFIG                       = jsonencode(local.ssm_housekeeper.config)
-      SERVICE_NAME                             = "ssm-housekeeper"
+      POWERTOOLS_SERVICE_NAME                  = "ssm-housekeeper"
       POWERTOOLS_TRACE_ENABLED                 = var.tracing_config.mode != null ? true : false
       POWERTOOLS_TRACER_CAPTURE_HTTPS_REQUESTS = var.tracing_config.capture_http_requests
       POWERTOOLS_TRACER_CAPTURE_ERROR          = var.tracing_config.capture_error